# lib/api-zod/src/generated/api/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get dashboard stats
 */
export const GetAdminStatsResponse = zod.object({
  totalRides: zod.number(),
  totalUsers: zod.number(),
  totalDrivers: zod.number(),
  totalEarnings: zod.number(),
  activeDrivers: zod.number(),
  ridesThisMonth: zod.number(),
  earningsThisMonth: zod.number(),
  avgRating: zod.number(),
});

/**
 * @summary List all users
 */
export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  status: zod.string(),
  totalRides: zod.number(),
  createdAt: zod.string(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary List all drivers
 */
export const ListDriversResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  vehicleType: zod.string(),
  vehicleNumber: zod.string(),
  rating: zod.number(),
  status: zod.string(),
  totalEarnings: zod.number(),
  totalRides: zod.number(),
  createdAt: zod.string(),
});
export const ListDriversResponse = zod.array(ListDriversResponseItem);

/**
 * @summary List all rides
 */
export const ListRidesQueryParams = zod.object({
  status: zod.coerce.string().optional(),
});

export const ListRidesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  driverId: zod.number().nullish(),
  userName: zod.string(),
  driverName: zod.string().nullish(),
  pickup: zod.string(),
  destination: zod.string(),
  vehicleType: zod.string(),
  rideMode: zod.string(),
  price: zod.number(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const ListRidesResponse = zod.array(ListRidesResponseItem);

/**
 * @summary Get recent rides
 */
export const GetRecentRidesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  driverId: zod.number().nullish(),
  userName: zod.string(),
  driverName: zod.string().nullish(),
  pickup: zod.string(),
  destination: zod.string(),
  vehicleType: zod.string(),
  rideMode: zod.string(),
  price: zod.number(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const GetRecentRidesResponse = zod.array(GetRecentRidesResponseItem);

/**
 * @summary Get daily analytics for chart
 */
export const GetDailyAnalyticsResponseItem = zod.object({
  date: zod.string(),
  rides: zod.number(),
  earnings: zod.number(),
  newUsers: zod.number(),
});
export const GetDailyAnalyticsResponse = zod.array(
  GetDailyAnalyticsResponseItem,
);

/**
 * @summary Assign a driver to a ride
 */
export const AssignDriverToRideParams = zod.object({
  id: zod.coerce.number(),
});

export const AssignDriverToRideBody = zod.object({
  driverId: zod.number(),
});

export const AssignDriverToRideResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  driverId: zod.number().nullish(),
  userName: zod.string(),
  driverName: zod.string().nullish(),
  pickup: zod.string(),
  destination: zod.string(),
  vehicleType: zod.string(),
  rideMode: zod.string(),
  price: zod.number(),
  status: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
  role: zod.string(),
});