# lib/api-zod/src/generated/types/driver.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Driver {
  id: number;
  name: string;
  email: string;
  phone: string;
  vehicleType: string;
  vehicleNumber: string;
  rating: number;
  status: string;
  totalEarnings: number;
  totalRides: number;
  createdAt: string;
}