# lib/api-zod/src/generated/types/ride.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Ride {
  id: number;
  userId: number;
  driverId?: number | null;
  userName: string;
  driverName?: string | null;
  pickup: string;
  destination: string;
  vehicleType: string;
  rideMode: string;
  price: number;
  status: string;
  createdAt: string;
}